backend/requirements.txt
websockets>=12.0
pystray>=0.19
pillow>=10.0
